+++ rio/debug/dev_tools/icons_page.py
@@ -330,7 +330,7 @@ Use the `rio.Icon` component like this:
                 rio.Text(
                     "Rio comes with a large array of icons out of the box. You can find them all here.",
                     justify="center",
-                    wrap=True,
+                    overflow="wrap",
                     style=rio.TextStyle(
                         font_size=1.1,
                         font_weight="bold",